refactor: more consistently type `auth_context` via `UserContextDict`
