+++ registry-pkgs/src/registry_pkgs/workflows/runner.py
@@ -66,6 +66,7 @@ from registry_pkgs.models.workflow import (
 )
 from registry_pkgs.telemetry.workflow_metrics import record_workflow_run
 from registry_pkgs.telemetry.workflow_tracing import trace_workflow_continuation, trace_workflow_run
+from registry_pkgs.types import UserContextDict
 from registry_pkgs.workflows.a2a_client import HeadersProvider
 from registry_pkgs.workflows.compiler import StepExecutor, compile_workflow, flatten_workflow_nodes
 from registry_pkgs.workflows.control import DirectiveQueue, WorkflowCancelledError
@@ -166,7 +167,7 @@ class WorkflowRunner:
         definition_id: str,
         user_text: str,
         *,
-        auth_context: dict[str, Any] | None,
+        auth_context: UserContextDict | None,
         existing_run_id: str,
         injected_outputs: dict[str, dict[str, Any]] | None = None,
         stop_after_node_id: str | None = None,
@@ -280,7 +281,7 @@ class WorkflowRunner:
     async def _build_registry(
         self,
         definition: WorkflowDefinition,
-        auth_context: dict[str, Any] | None,
+        auth_context: UserContextDict | None,
     ) -> dict[str, StepExecutor]:
         """Extract executor keys + pool nodes from the definition and resolve them.
 
@@ -322,7 +323,7 @@ class WorkflowRunner:
         self,
         *,
         existing_run_id: str,
-        auth_context: dict[str, Any] | None,
+        auth_context: UserContextDict | None,
     ) -> tuple[WorkflowRun, list[NodeRun]]:
         """Resume a run that is holding at one or more pending requirements.
