+++ registry/src/registry/auth/dependencies.py
@@ -1,48 +1,18 @@
 import logging
-from typing import Annotated, TypedDict
+from typing import Annotated
 
 from fastapi import Depends, HTTPException, Request, status
 from itsdangerous import URLSafeTimedSerializer
+
+from registry_pkgs.types import UserContextDict
 
 from ..core.config import settings
 
 logger = logging.getLogger(__name__)
 
-
-class UserContextDict(TypedDict):
-    """
-    UserContextDict is the type of the dictionary set as the Request.state.user attribute for each incoming request
-    **to a non-public route** by the UnifiedAuthMiddlware.
-    If a FastAPI path operation function needs to access this UserContextDict information, it can retrieve this dictionary
-    by asking for the `user_context: CurrentUser` dependency injection.
-    If an MCP tool/resource/prompt handler function needs it, it can retrieve this dictionary by asking for the
-    `ctx: Context` dependency injection (`mcp.server.fastmcp.Context`) and then accessing
-    `ctx.request_context.request.state.user`.
-    """
-
-    # From the "user_id" field of the JWT claim.
-    user_id: str | None
-
-    # From the "client_id" claim of the JWT.
-    client_id: str
-
-    # From the "sub" field of the JWT claim.
-    username: str | None
-
-    # From the "groups" field of the JWT claim.
-    groups: list[str]
-
-    # Converted from the "scope" field of the JWT claim, or mapped from the "groups" field if "scope" doesn't exist.
-    scopes: list[str]
-
-    # "jwt" if the JWT token comes from the Authorization header. "traditional" if from browser cookie.
-    auth_method: str
-
-    # "jwt" if the JWT token comes from the Authorization header. "local" if from browser cookie.
-    provider: str
-
-    # "jwt_auth" if the JWT token comes from the Authorization header. "jwt_session_auth" if from browser cookie.
-    auth_source: str
+# UserContextDict is re-exported here (defined in registry_pkgs.types) so existing
+# `from registry.auth.dependencies import UserContextDict` call sites keep working, and so
+# registry_pkgs code — which cannot import from registry — can use the same type.
 
 
 def get_current_user(request: Request) -> UserContextDict:
